requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.0
distlib==0.3.7
filelock==3.12.4
Flask==3.0.0
fonttools==4.43.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.24.2
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
platformdirs==3.10.0
pydocstyle==6.3.0
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3.post1
requests==2.31.0
scipy==1.11.3
six==1.16.0
snowballstemmer==2.2.0
tzdata==2023.3
urllib3==2.0.6
virtualenv==20.24.5
Werkzeug==3.0.0
